+++ matjalalzz/.gitignore
@@ -36,5 +36,3 @@ out/
 ### VS Code ###
 .vscode/
 .env
-
-.env
