+++ matjalalzz/src/main/java/shop/matjalalzz/comment/dao/CommentRepository.java
@@ -1,0 +1,12 @@
+package shop.matjalalzz.comment.dao;
+
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import shop.matjalalzz.comment.entity.Comment;
+
+@Repository
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+
+    List<Comment> findAllByPartyId(Long partyId);
+}
