+++ matjalalzz/src/main/java/shop/matjalalzz/comment/dto/CommentResponse.java
@@ -1,0 +1,19 @@
+package shop.matjalalzz.comment.dto;
+
+import java.time.LocalDateTime;
+import lombok.Builder;
+
+@Builder
+public record CommentResponse(
+    Long commentId,
+    Long parentId,
+    String content,
+    LocalDateTime createdAt,
+    Writer writer) {
+
+    @Builder
+    public record Writer(Long userId, String nickname) {
+
+    }
+}
+
