--- matjalalzz/src/main/java/shop/matjalalzz/domain/comment/dao/CommentRepository.java
@@ -1,12 +1,0 @@
-package shop.matjalalzz.domain.comment.dao;
-
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import shop.matjalalzz.domain.comment.entity.Comment;
-
-@Repository
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-
-    List<Comment> findAllByPartyId(Long partyId);
-}
