--- matjalalzz/src/main/java/shop/matjalalzz/domain/review/dao/ReviewRepository.java
@@ -1,17 +1,0 @@
-package shop.matjalalzz.domain.review.dao;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
-import shop.matjalalzz.domain.review.entity.Review;
-
-@Repository
-public interface ReviewRepository extends JpaRepository<Review, Long> {
-
-    @Query("SELECT r FROM Review r WHERE r.shop.id = :shopId AND (r.id < :cursor OR :cursor = 0) ORDER BY r.id DESC")
-    Page<Review> findByShopIdAndCursor(@Param("shopId") Long shopId, @Param("cursor") Long cursor,
-        Pageable pageable);
-}
