--- matjalalzz/src/main/java/shop/matjalalzz/domain/review/dto/ReviewPageResponse.java
@@ -1,13 +1,0 @@
-package shop.matjalalzz.domain.review.dto;
-
-import java.util.List;
-import lombok.Builder;
-
-@Builder
-public record ReviewPageResponse(
-    Long nextCursor,
-    List<ReviewResponse> reviews
-) {
-
-
-}
