+++ matjalalzz/src/main/java/shop/matjalalzz/review/dto/ReviewPageResponse.java
@@ -1,0 +1,13 @@
+package shop.matjalalzz.review.dto;
+
+import java.util.List;
+import lombok.Builder;
+
+@Builder
+public record ReviewPageResponse(
+    Long nextCursor,
+    List<ReviewResponse> reviews
+) {
+
+
+}
