+++ matjalalzz/src/main/java/shop/matjalalzz/review/dto/ReviewResponse.java
@@ -1,0 +1,16 @@
+package shop.matjalalzz.review.dto;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import lombok.Builder;
+
+@Builder
+public record ReviewResponse(
+    Long reviewId,
+    String userNickname,
+    Double rating,
+    String content,
+    LocalDateTime createdAt,
+    List<String> images) {
+
+}
